feat(FUSERS_LIST): Show groups collapsed by default and memorize groups' states (refs #5441)
The group's tree is displayed by default with all groups collapsed.

When opening or closing a group, the group's state is memorized in order
to redisplay the tree's state on subsequent views.

The tree's state is memorized in the FUSERS_GTREESTATE user's parameter.

The interface also provides a global collapse and expand button as in
the USERCARD:CHOOSEGROUP interface.